# openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorDto": {
        "properties": {
          "error": {
            "description": "Short error identifier. Usually the standard reason phrase; domain errors use a code, e.g. step_up_required.",
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "description": "Human-readable message. An array when request validation failed, one entry per constraint.",
            "example": "Please re-verify your identity to continue.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ]
          },
          "statusCode": {
            "description": "HTTP status code, repeated in the body.",
            "example": 400,
            "type": "integer"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
      "AppleLoginDto": {
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "fullName": {
            "description": "User's full name (Apple only sends this on first sign-in)",
            "type": "string"
          },
          "identityToken": {
            "description": "Apple identity token (JWT) from the native sign-in flow",
            "type": "string"
          }
        },
        "required": [
          "identityToken"
        ],
        "type": "object"
      },
      "AuthResponseDto": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresIn": {
            "description": "Access-token lifetime in seconds",
            "type": "integer"
          },
          "refreshToken": {
            "type": "string"
          },
          "tokenType": {
            "example": "Bearer",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn",
          "user"
        ],
        "type": "object"
      },
...
      "GoogleLoginDto": {
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "idToken": {
            "description": "Google ID token from the native sign-in flow",
            "type": "string"
          }
        },
        "required": [
          "idToken"
        ],
        "type": "object"
      },
...
      "LogoutDto": {
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "type": "object"
      },
...
      "OtpChallengeResponseDto": {
        "properties": {
          "devCode": {
            "description": "Dev/testing only — present when AUTH_EXPOSE_OTP is enabled (non-production).",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "resendAvailableAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "expiresAt",
          "resendAvailableAt"
        ],
        "type": "object"
      },
...
      "RefreshDto": {
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "type": "object"
      },
...
      "RequestEmailDto": {
        "properties": {
          "email": {
            "example": "leo@example.com",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
      "RequestOtpDto": {
        "properties": {
          "deviceId": {
            "description": "Opaque device identifier for this session",
            "type": "string"
          },
          "phone": {
            "description": "Phone in E.164 format",
            "example": "+4915123456789",
            "type": "string"
          }
        },
        "required": [
          "phone"
        ],
        "type": "object"
      },
...
      "UserDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "displayName": {
            "nullable": true,
            "type": "string"
          },
          "email": {
            "nullable": true,
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "phone": {
            "nullable": true,
            "type": "string"
          },
          "phoneVerified": {
            "type": "boolean"
          },
          "status": {
            "enum": [
              "registered",
              "active",
              "restricted",
              "suspended",
              "deletion_pending",
              "deleted"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "phone",
          "email",
          "displayName",
          "status",
          "phoneVerified",
          "emailVerified",
          "createdAt"
        ],
        "type": "object"
      },
...
      "VerifyEmailDto": {
        "properties": {
          "code": {
            "description": "The 6-digit code sent by email",
            "example": "123456",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
      "VerifyOtpDto": {
        "properties": {
          "code": {
            "example": "482159",
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "phone": {
            "description": "Phone in E.164 format",
            "example": "+4915123456789",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "code"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Backend API for ShowUp — real-world dating. Source of truth for users, profiles, check-ins, matching, dates, payments, notifications, analytics, safety and admin.",
    "title": "ShowUp Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/account/delete-request": {
      "post": {
        "operationId": "requestAccountDeletion",
        "parameters": [],
        "responses": {
          "202": {
            "description": "Deletion requested; no response body."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "account"
        ]
      }
    },
    "/account/delete-request/cancel": {
      "post": {
        "operationId": "cancelAccountDeletion",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Deletion request cancelled; no response body."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "account"
        ]
      }
    },
...
    "/auth/apple": {
      "post": {
        "operationId": "loginWithApple",
        "parameters": [
          {
            "in": "header",
            "name": "user-agent",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleLoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/email/start": {
      "post": {
        "operationId": "startEmailVerification",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestEmailDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpChallengeResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/email/verify": {
      "post": {
        "operationId": "verifyEmail",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/google": {
      "post": {
        "operationId": "loginWithGoogle",
        "parameters": [
          {
            "in": "header",
            "name": "user-agent",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "account"
        ]
      }
    },
    "/auth/phone/start": {
      "post": {
        "operationId": "startPhoneVerification",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpChallengeResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/phone/verify": {
      "post": {
        "operationId": "verifyPhone",
        "parameters": [
          {
            "in": "header",
            "name": "user-agent",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "refreshAuthToken",
        "parameters": [
          {
            "in": "header",
            "name": "user-agent",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Unexpected server error."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "servers": [],
...
}